refactor: reorganize draft hooks and components for improved structure and clarity
