+++ packages/web/src/views/Calendar/components/Event/Grid/GridEventPreview/GridEventPreview.tsx
@@ -15,7 +15,7 @@ import { WeekProps } from "@web/views/Calendar/hooks/useWeek";
 import { DateCalcs } from "@web/views/Calendar/hooks/grid/useDateCalcs";
 import { SpaceCharacter } from "@web/components/SpaceCharacter";
 import { snapToGrid } from "@web/views/Calendar/components/Event/Grid/GridEventPreview/snap.grid";
-import { MouseCoords } from "@web/views/Calendar/hooks/draft/useMousePosition";
+import { MouseCoords } from "@web/views/Calendar/hooks/mouse/useMousePosition";
 
 import { getItemStyles, layerStyles, StyledGridEventPreview } from "./styled";
